# [arm64] JIT: Enable CSE/hoisting for "arrayBase + elementOffset"