--- a/src/coreclr/jit/codegenlinear.cpp
+++ b/src/coreclr/jit/codegenlinear.cpp
@@ -1574,12 +1574,14 @@ void CodeGen::genConsumeRegs(GenTree* tree)
         {
             genConsumeAddress(tree);
         }
+#ifdef TARGET_ARM64
         else if (tree->OperIs(GT_BFIZ))
         {
             GenTreeCast* cast = tree->gtGetOp1()->AsCast();
             assert(cast->isContained());
             genConsumeAddress(cast->CastOp());
         }
+#endif
         else if (tree->OperIsLocalRead())
         {
             // A contained lcl var must be living on stack and marked as reg optional, or not be a