--- a/src/coreclr/jit/codegenlinear.cpp
+++ b/src/coreclr/jit/codegenlinear.cpp
@@ -1577,6 +1577,7 @@ void CodeGen::genConsumeRegs(GenTree* tree)
 #ifdef TARGET_ARM64
         else if (tree->OperIs(GT_BFIZ))
         {
+            // Can be contained as part of LEA on ARM64
             GenTreeCast* cast = tree->gtGetOp1()->AsCast();
             assert(cast->isContained());
             genConsumeAddress(cast->CastOp());