Closes https://github.com/dotnet/runtime/issues/35618, https://github.com/dotnet/runtime/issues/34810

Usually, for a typical array access we create the following tree
```
arrRef + (index + elemOffset)
```
which is basically `invariant1 + (X + invariant2)`, if we change it to:
```
(arrRef + elemOffset) + index
```
_I assume it's safe from GC's point of view, both ADDs are byref and within the objects_ 

we'll be able to hoist/CSE the invariant part `arrRef + elemOffset` on arm/arm64. We don't really need it on XArch because it will be handled by a more powerful addressing mode such as `mov      r9d, dword ptr [rdx+4*r9+16]`.
So, essentially, this PR changes morphing of GT_INDEX like this:
  
![image](https://user-images.githubusercontent.com/523221/140639712-63f1f127-d633-4939-adc5-13269fddfb79.png)

**UPD** Also, this PR folds LEA(BFIZ) into load/store with scale and sign/zero extension.

Example:
```csharp
int Test(int[] array, int i, int j) 
    => array[i] + array[j];

int Sum(int[] array)
{
    int sum = 0;
    foreach (int item in array)
        sum += item;
    return sum;
}
```
codegen diff:
```diff
; Method Tests:Test(System.Int32[],int,int):int:this
G_M15096_IG01:
            stp     fp, lr, [sp,#-16]!
            mov     fp, sp
G_M15096_IG02:
            ldr     w0, [x1,#8]
            cmp     w2, w0
            bhs     G_M15096_IG04
-           ubfiz   x2, x2, #2, #32
            add     x2, x2, #16
-           ldr     w2, [x1, x2]
+           ldr     w2, [x1, w2, UXTW #2]
            cmp     w3, w0
            bhs     G_M15096_IG04
-           ubfiz   x0, x3, #2, #32
-           add     x0, x0, #16     ;; CSE'd
-           ldr     w0, [x1, x0]
+           ldr     w0, [x1, w3, UXTW #2]
            add     w0, w2, w0
G_M15096_IG03:
            ldp     fp, lr, [sp],#16
            ret     lr
G_M15096_IG04:
            bl      CORINFO_HELP_RNGCHKFAIL
            bkpt    
-; Total bytes of code: 72
+; Total bytes of code: 60


; Method Tests:Sum(System.Int32[]):int:this
G_M56165_IG01:
            stp     fp, lr, [sp,#-16]!
            mov     fp, sp
G_M56165_IG02:
            mov     w0, wzr
            mov     w2, wzr
            ldr     w3, [x1,#8]
            cmp     w3, #0
            ble     G_M56165_IG04
+           add     x1, x1, #16    ;; 'baseRef + elemOffset' is hoisted
G_M56165_IG03:
-           ubfiz   x4, x2, #2, #32
-           add     x4, x4, #16
-           ldr     w4, [x1, x4]    ;; -2 4byte instructions from the loop body, it has now a perfect codegen
+           ldr     w4, [x1, w2, UXTW #2]
            add     w0, w0, w4
            add     w2, w2, #1
            cmp     w3, w2
            bgt     G_M56165_IG03
G_M56165_IG04:
            ldp     fp, lr, [sp],#16
            ret     lr
-; Total bytes of code: 64
+; Total bytes of code: 60


```
Benchmark: "Sum all elements in an array":
```csharp
[Benchmark]
[ArgumentsSource(nameof(TestData))]
public int Sum(int[] array)
{
    int sum = 0;
    foreach (int t in array)
        sum += t;
    return sum;
}
```
Benchmark results on Apple M1 arm64:
```
                             | Method |         array |     Mean |    Error |
                             |------- |-------------- |---------:|---------:|
Main without #61045 + #61035 |    Sum | Int32[100000] | 60,08 us | 0.015 us | .NET 6.0 codegen
                Current Main |    Sum | Int32[100000] | 53.78 us | 0.007 us |
                     This PR |    Sum | Int32[100000] | 32.61 us | 0.082 us | ~2x faster than .NET 6.0 here
```
_^ tested with and without loop-alignment_

Codegen diff:
- Main before #61045 and #61035 (basically, it's .NET 6.0 codegen) vs This PR https://www.diffchecker.com/smwmMHvi

Diffs are big but in both - and + directions:

## benchmarks.run.windows.arm64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 7603860 (overridden on cmd)
Total bytes of diff: 7582960 (overridden on cmd)
Total bytes of delta: -20900 (-0.27 % of base)
    diff is an improvement.
    relative diff is an improvement.
```
<details>

<summary>Detail diffs</summary>

```


Top file regressions (bytes):
          48 : 1745.dasm (2.82% of base)
          48 : 3317.dasm (2.86% of base)
          48 : 6982.dasm (2.12% of base)
          36 : 829.dasm (3.00% of base)
          32 : 19120.dasm (8.33% of base)
          28 : 8405.dasm (2.06% of base)
          24 : 21759.dasm (1.59% of base)
          24 : 18601.dasm (7.79% of base)
          24 : 6662.dasm (6.00% of base)
          24 : 14412.dasm (1.42% of base)
          24 : 8565.dasm (2.90% of base)
          20 : 1037.dasm (4.46% of base)
          16 : 11927.dasm (3.10% of base)
          16 : 2104.dasm (1.38% of base)
          16 : 416.dasm (2.35% of base)
          16 : 6641.dasm (1.98% of base)
          16 : 1734.dasm (2.20% of base)
          12 : 12997.dasm (0.54% of base)
          12 : 16129.dasm (5.88% of base)
          12 : 3046.dasm (3.37% of base)

Top file improvements (bytes):
        -660 : 3506.dasm (-6.48% of base)
        -520 : 19287.dasm (-8.86% of base)
        -512 : 12906.dasm (-12.77% of base)
        -508 : 23202.dasm (-8.34% of base)
        -504 : 15238.dasm (-8.21% of base)
        -500 : 13550.dasm (-2.09% of base)
        -484 : 21964.dasm (-8.85% of base)
        -436 : 14847.dasm (-11.83% of base)
        -344 : 2494.dasm (-1.85% of base)
        -324 : 1245.dasm (-4.29% of base)
        -324 : 2715.dasm (-3.20% of base)
        -324 : 11506.dasm (-3.20% of base)
        -248 : 19341.dasm (-4.62% of base)
        -248 : 21403.dasm (-2.68% of base)
        -244 : 6643.dasm (-1.40% of base)
        -240 : 12513.dasm (-5.37% of base)
        -236 : 2714.dasm (-2.38% of base)
        -236 : 22578.dasm (-2.41% of base)
        -220 : 18295.dasm (-3.24% of base)
        -220 : 11505.dasm (-3.13% of base)

1220 total files with Code Size differences (1136 improved, 84 regressed), 368 unchanged.

Top method regressions (bytes):
          48 ( 2.82% of base) : 1745.dasm - System.Reflection.Emit.MethodBuilder:CreateMethodBodyHelper(System.Reflection.Emit.ILGenerator):this
          48 ( 2.12% of base) : 6982.dasm - System.Runtime.Serialization.Formatters.Binary.ObjectWriter:WriteArray(System.Runtime.Serialization.Formatters.Binary.WriteObjectInfo,System.Runtime.Serialization.Formatters.Binary.NameInfo,System.Runtime.Serialization.Formatters.Binary.WriteObjectInfo):this
          48 ( 2.86% of base) : 3317.dasm - System.Xml.XmlConvert:EncodeName(System.String,bool,bool):System.String
          36 ( 3.00% of base) : 829.dasm - System.Sha1ForNonSecretPurposes:Finish(System.Span`1[Byte]):this
          32 ( 8.33% of base) : 19120.dasm - System.Collections.Concurrent.ConcurrentBag`1[Int32][System.Int32]:ToArray():System.Int32[]:this
          28 ( 2.06% of base) : 8405.dasm - System.Diagnostics.StackFrameHelper:InitializeSourceInfo(int,bool,System.Exception):this
          24 ( 7.79% of base) : 18601.dasm - Benchstone.BenchI.AddArray:Test():bool:this
          24 ( 1.42% of base) : 14412.dasm - MessagePack.Formatters.MicroBenchmarks_Serializers_CollectionsOfPrimitivesFormatter1:Deserialize(System.Byte[],int,MessagePack.IFormatterResolver,byref):MicroBenchmarks.Serializers.CollectionsOfPrimitives:this
          24 ( 1.59% of base) : 21759.dasm - System.Buffers.Text.Utf8Formatter:TryFormat(System.Decimal,System.Span`1[Byte],byref,System.Buffers.StandardFormat):bool
          24 ( 6.00% of base) : 6662.dasm - System.Text.RegularExpressions.RegexCompiler:Goto(int):this
          24 ( 2.90% of base) : 8565.dasm - Utf8Json.JsonSerializer:Deserialize(System.IO.Stream,Utf8Json.IJsonFormatterResolver):System.__Canon
          20 ( 4.46% of base) : 1037.dasm - System.TimeZoneInfo:TryConvertIanaIdToWindowsId(System.String,bool,byref):bool
          16 ( 3.10% of base) : 11927.dasm - Benchstone.BenchI.XposMatrix:Inner(System.Int32[][],int)
          16 ( 1.38% of base) : 2104.dasm - System.DefaultBinder:FindMostSpecific(System.Reflection.ParameterInfo[],System.Int32[],System.Type,System.Reflection.ParameterInfo[],System.Int32[],System.Type,System.Type[],System.Object[]):int
          16 ( 2.35% of base) : 416.dasm - System.Reflection.Associates:AssignAssociates(int,System.RuntimeType,System.RuntimeType):System.Reflection.RuntimeMethodInfo
          16 ( 2.20% of base) : 1734.dasm - System.Reflection.Emit.__ExceptionInfo:MarkHelper(int,int,System.Type,int):this
          16 ( 1.98% of base) : 6641.dasm - System.Text.RegularExpressions.RegexCompiler:GenerateForwardSection():this
          12 ( 1.86% of base) : 20019.dasm - Benchstone.BenchI.Array2:Bench(int):bool
          12 ( 2.11% of base) : 20022.dasm - Benchstone.BenchI.Array2:VerifyCopy(System.Int32[][][],System.Int32[][][]):bool
          12 ( 1.03% of base) : 10242.dasm - EMFloat:AddSubInternalFPF(ubyte,byref,byref,byref)

Top method improvements (bytes):
        -660 (-6.48% of base) : 3506.dasm - System.Text.RegularExpressions.RegexInterpreter:Go():this
        -520 (-8.86% of base) : 19287.dasm - DynamicClass:Regex4_Go(System.Text.RegularExpressions.RegexRunner)
        -512 (-12.77% of base) : 12906.dasm - DynamicClass:Regex3_Go(System.Text.RegularExpressions.RegexRunner)
        -508 (-8.34% of base) : 23202.dasm - DynamicClass:Regex1_Go(System.Text.RegularExpressions.RegexRunner)
        -504 (-8.21% of base) : 15238.dasm - DynamicClass:Regex4_Go(System.Text.RegularExpressions.RegexRunner)
        -500 (-2.09% of base) : 13550.dasm - DynamicClass:_DynamicMethod0(System.IO.TextWriter,MicroBenchmarks.Serializers.MyEventsListerItem,int)
        -484 (-8.85% of base) : 21964.dasm - DynamicClass:Regex1_Go(System.Text.RegularExpressions.RegexRunner)
        -436 (-11.83% of base) : 14847.dasm - DynamicClass:Regex3_Go(System.Text.RegularExpressions.RegexRunner)
        -344 (-1.85% of base) : 2494.dasm - DynamicClass:_DynamicMethod0(System.IO.TextWriter,MicroBenchmarks.Serializers.IndexViewModel,int)
        -324 (-3.20% of base) : 11506.dasm - Jil.Deserialize.Methods:ParseISO8601Date(System.IO.TextReader,System.Char[],int,int):System.DateTime
        -324 (-3.20% of base) : 2715.dasm - Jil.Deserialize.Methods:ParseISO8601DateThunkReader(byref,System.Char[],int,int):System.DateTime
        -324 (-4.29% of base) : 1245.dasm - System.Text.RegularExpressions.RegexWriter:EmitFragment(int,System.Text.RegularExpressions.RegexNode,int):this
        -248 (-2.68% of base) : 21403.dasm - DynamicClass:Regex2_Go(System.Text.RegularExpressions.RegexRunner)
        -248 (-4.62% of base) : 19341.dasm - Utf8Json.Formatters.ISO8601DateTimeFormatter:Deserialize(byref,Utf8Json.IJsonFormatterResolver):System.DateTime:this
        -244 (-1.40% of base) : 6643.dasm - System.Text.RegularExpressions.RegexCompiler:GenerateOneCode():this
        -240 (-5.37% of base) : 12513.dasm - Utf8Json.Formatters.ISO8601DateTimeOffsetFormatter:Deserialize(byref,Utf8Json.IJsonFormatterResolver):System.DateTimeOffset:this
        -236 (-2.41% of base) : 22578.dasm - Jil.Deserialize.Methods:_ReadISO8601DateThunkReader(byref,System.Char[]):System.DateTime
        -236 (-2.38% of base) : 2714.dasm - Jil.Deserialize.Methods:_ReadISO8601DateWithOffsetThunkReader(byref,System.Char[]):System.DateTimeOffset
        -220 (-3.24% of base) : 18295.dasm - Jil.Deserialize.Methods:_ReadISO8601Date(System.IO.TextReader,System.Char[]):System.DateTime
        -220 (-3.13% of base) : 11505.dasm - Jil.Deserialize.Methods:_ReadISO8601DateWithOffset(System.IO.TextReader,System.Char[]):System.DateTimeOffset

Top method regressions (percentages):
          32 ( 8.33% of base) : 19120.dasm - System.Collections.Concurrent.ConcurrentBag`1[Int32][System.Int32]:ToArray():System.Int32[]:this
          24 ( 7.79% of base) : 18601.dasm - Benchstone.BenchI.AddArray:Test():bool:this
          12 ( 7.50% of base) : 14104.dasm - System.Collections.CreateAddAndClear`1[Int32][System.Int32]:Array():System.Int32[]:this
          24 ( 6.00% of base) : 6662.dasm - System.Text.RegularExpressions.RegexCompiler:Goto(int):this
          12 ( 5.88% of base) : 16129.dasm - Microsoft.CodeAnalysis.Collections.ByteSequenceComparer:Equals(System.Collections.Immutable.ImmutableArray`1[Byte],System.Collections.Immutable.ImmutableArray`1[Byte]):bool
          12 ( 5.88% of base) : 17988.dasm - System.Reflection.Internal.ByteSequenceComparer:Equals(System.Collections.Immutable.ImmutableArray`1[Byte],System.Collections.Immutable.ImmutableArray`1[Byte]):bool
          12 ( 5.17% of base) : 5864.dasm - SslCredKey:Equals(SslCredKey):bool:this
          12 ( 4.48% of base) : 5945.dasm - System.Security.Cryptography.Primitives.Tests.Performance.Perf_FixedTimeEquals:Setup(System.String,System.String):this
          20 ( 4.46% of base) : 1037.dasm - System.TimeZoneInfo:TryConvertIanaIdToWindowsId(System.String,bool,byref):bool
           8 ( 4.35% of base) : 6659.dasm - System.Text.RegularExpressions.RegexCompiler:TrackUnique(int):this
           8 ( 4.00% of base) : 9847.dasm - System.Reflection.Metadata.MetadataReader:CombineRowCounts(System.Int32[],System.Int32[],ubyte):System.Int32[]
          12 ( 4.00% of base) : 2564.dasm - System.RuntimeTypeHandle:CopyRuntimeTypeHandles(System.Type[],byref):System.IntPtr[]
          12 ( 3.85% of base) : 9862.dasm - System.Diagnostics.StackFrame:.ctor(System.Diagnostics.StackFrameHelper,int,bool):this
           8 ( 3.77% of base) : 4816.dasm - ListPartition`1[Char][System.Char]:ToArray():System.Char[]:this
          12 ( 3.70% of base) : 6980.dasm - System.Collections.HashHelpers:GetPrime(int):int
          12 ( 3.70% of base) : 6.dasm - System.Collections.HashHelpers:GetPrime(int):int
           4 ( 3.57% of base) : 10237.dasm - EMFloat:memmove(byref,byref)
          12 ( 3.37% of base) : 3046.dasm - Newtonsoft.Json.DefaultJsonNameTable:Add(System.String):System.String:this
           4 ( 3.33% of base) : 11430.dasm - EMFloatClass:memmove(InternalFPF,InternalFPF)
           8 ( 3.33% of base) : 20974.dasm - SelectIListIterator`2[Int32,Int32][System.Int32,System.Int32]:ToArray():System.Int32[]:this

Top method improvements (percentages):
         -60 (-20.00% of base) : 20833.dasm - Benchstone.BenchI.BubbleSort2:Test():bool:this
         -20 (-17.24% of base) : 12447.dasm - System.Text.RegularExpressions.RegexInterpreter:TrackPush(int,int):this
        -176 (-17.12% of base) : 13511.dasm - IDEAEncryption:de_key_idea(System.Char[],System.Char[])
         -20 (-16.67% of base) : 14479.dasm - System.Text.RegularExpressions.RegexInterpreter:TrackPush2(int,int):this
         -92 (-15.75% of base) : 6845.dasm - Benchstone.BenchI.HeapSort:Inner(System.Int32[],int)
         -32 (-15.69% of base) : 8335.dasm - V8.Crypto.Arcfour:next():int:this
         -52 (-15.48% of base) : 14989.dasm - Span.Sorting:TestQuickSortArray(System.Int32[],int,int)
         -24 (-15.00% of base) : 2375.dasm - <>c__DisplayClass14_0:<_ReadFieldOperands>b__0(int):int:this
         -40 (-13.70% of base) : 14518.dasm - Parser:Parse4Dig(int,byref):bool:this
         -40 (-13.51% of base) : 6150.dasm - Newtonsoft.Json.Utilities.DateTimeParser:Parse4Digit(int,byref):bool:this
         -24 (-13.33% of base) : 19671.dasm - Benchstone.BenchI.Array1:VerifySort(System.Int32[]):bool
         -12 (-13.04% of base) : 3510.dasm - System.Text.RegularExpressions.RegexInterpreter:TrackPush(int):this
        -512 (-12.77% of base) : 12906.dasm - DynamicClass:Regex3_Go(System.Text.RegularExpressions.RegexRunner)
         -24 (-12.50% of base) : 9549.dasm - System.Net.Internals.SocketAddress:CopyAddressSizeIntoBuffer():this
         -12 (-12.50% of base) : 15559.dasm - System.Text.RegularExpressions.RegexInterpreter:TrackPush2(int):this
         -12 (-12.50% of base) : 7516.dasm - System.Xml.XmlException:BuildCharExceptionArgs(System.String,int):System.String[]
         -24 (-12.24% of base) : 20517.dasm - Benchstone.BenchI.BubbleSort:VerifySort(System.Int32[],int):bool
        -436 (-11.83% of base) : 14847.dasm - DynamicClass:Regex3_Go(System.Text.RegularExpressions.RegexRunner)
         -52 (-11.50% of base) : 11618.dasm - System.SpanHelpers:LastIndexOf(byref,int,int):int
         -52 (-11.21% of base) : 10989.dasm - System.SpanHelpers:LastIndexOf(byref,ushort,int):int

1220 total methods with Code Size differences (1136 improved, 84 regressed), 368 unchanged.

```

</details>

--------------------------------------------------------------------------------

## coreclr_tests.pmi.windows.arm64.checked.mch:

```

Summary of Code Size diffs:
(Lower is better)

Total bytes of base: 165345900 (overridden on cmd)
Total bytes of diff: 165319844 (overridden on cmd)
Total bytes of delta: -26056 (-0.02 % of base)
    diff is an improvement.
    relative diff is an improvement.
```
<details>

<summary>Detail diffs</summary>

```


Top file regressions (bytes):
         104 : 239428.dasm (2.61% of base)
          68 : 167158.dasm (9.04% of base)
          68 : 100367.dasm (8.99% of base)
          68 : 101273.dasm (9.04% of base)
          68 : 166934.dasm (9.09% of base)
          68 : 166744.dasm (9.04% of base)
          68 : 167347.dasm (9.09% of base)
          68 : 100446.dasm (9.04% of base)
          68 : 101196.dasm (8.99% of base)
          68 : 166823.dasm (9.09% of base)
          68 : 101385.dasm (9.04% of base)
          68 : 100557.dasm (9.04% of base)
          68 : 167235.dasm (9.09% of base)
          64 : 100560.dasm (10.74% of base)
          64 : 101356.dasm (10.74% of base)
          64 : 101433.dasm (10.74% of base)
          64 : 166952.dasm (10.74% of base)
          64 : 166965.dasm (8.51% of base)
          64 : 166968.dasm (8.51% of base)
          64 : 100465.dasm (8.33% of base)

Top file improvements (bytes):
       -1548 : 242531.dasm (-14.79% of base)
        -664 : 94.dasm (-6.53% of base)
        -560 : 234600.dasm (-3.52% of base)
        -556 : 234597.dasm (-3.50% of base)
        -544 : 189301.dasm (-18.63% of base)
        -544 : 189050.dasm (-18.63% of base)
        -544 : 188745.dasm (-18.63% of base)
        -448 : 189025.dasm (-17.55% of base)
        -448 : 189276.dasm (-17.55% of base)
        -448 : 188720.dasm (-17.55% of base)
        -324 : 62.dasm (-4.29% of base)
        -300 : 189046.dasm (-14.85% of base)
        -300 : 189297.dasm (-14.85% of base)
        -300 : 188741.dasm (-14.85% of base)
        -252 : 189024.dasm (-13.82% of base)
        -252 : 189275.dasm (-13.82% of base)
        -252 : 188719.dasm (-13.82% of base)
        -196 : 189049.dasm (-18.15% of base)
        -196 : 189300.dasm (-18.15% of base)
        -196 : 188744.dasm (-18.15% of base)

5575 total files with Code Size differences (4500 improved, 1075 regressed), 1309 unchanged.

Top method regressions (bytes):
         104 ( 2.61% of base) : 239428.dasm - Co1245GetBytes_double:runTest():bool:this
          68 ( 9.09% of base) : 166934.dasm - testout1:Sub_Funclet_207():int
          68 ( 9.09% of base) : 167347.dasm - testout1:Sub_Funclet_207():int
          68 ( 9.04% of base) : 101385.dasm - testout1:Sub_Funclet_207():int
          68 ( 9.04% of base) : 100557.dasm - testout1:Sub_Funclet_207():int
          68 ( 9.04% of base) : 101273.dasm - testout1:Sub_Funclet_287():int
          68 ( 9.04% of base) : 100446.dasm - testout1:Sub_Funclet_287():int
          68 ( 9.09% of base) : 166823.dasm - testout1:Sub_Funclet_287():int
          68 ( 9.09% of base) : 167235.dasm - testout1:Sub_Funclet_287():int
          68 ( 9.04% of base) : 167158.dasm - testout1:Sub_Funclet_399():int
          68 ( 8.99% of base) : 100367.dasm - testout1:Sub_Funclet_399():int
…