@EgorBo 2021-11-07
I ran a couple of dotnet/performance benchmarks:
```
| Faster                                                          | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| Benchstone.BenchI.TreeInsert.Test                               |      1.56 |         47667.43 |         30543.60 | bimodal |
| SeekUnroll.Test(boxedIndex: 27)                                 |      1.50 |     938645542.00 |     625722750.00 |         |
| SeekUnroll.Test(boxedIndex: 3)                                  |      1.50 |     938589521.00 |     625770375.00 |         |
| SciMark2.kernel.benchSparseMult                                 |      1.15 |     469878458.00 |     407268833.00 |         |
| System.Numerics.Tests.Perf_BitOperations.LeadingZeroCount_uint  |      1.15 |           597.57 |           520.92 |         |
| System.Numerics.Tests.Perf_BitOperations.LeadingZeroCount_ulong |      1.14 |           598.13 |           522.41 |         |
| Benchstone.BenchI.CSieve.Test                                   |      1.14 |       4139455.41 |       3623725.51 |         |
| Benchstone.BenchI.AddArray2.Test                                |      1.12 |       5261706.59 |       4687558.60 |         |
| Burgers.Test1                                                   |      1.10 |     138233916.50 |     125235312.50 |         |
| Burgers.Test2                                                   |      1.09 |     137031989.50 |     125238250.00 |         |
| System.Numerics.Tests.Perf_BitOperations.PopCount_uint          |      1.08 |           515.04 |           475.31 |         |
| System.Numerics.Tests.Perf_BitOperations.Log2_ulong             |      1.08 |           609.98 |           563.30 |         |
| System.Numerics.Tests.Perf_BitOperations.Log2_uint              |      1.08 |           608.38 |           562.71 |         |
| SciMark2.kernel.benchFFT                                        |      1.07 |     520232666.00 |     485479459.00 |         |
| ByteMark.BenchNumericSortJagged                                 |      1.07 |    1061907479.50 |     994064208.00 |         |
| Burgers.Test0                                                   |      1.06 |     174896583.00 |     164355125.00 |         |
| Benchstone.BenchI.XposMatrix.Test                               |      1.06 |         12267.20 |         11569.69 |         |
| SciMark2.kernel.benchmarkLU                                     |      1.05 |     467178750.00 |     443115708.00 |         |
```