@EgorBo 2021-11-07
`CopyArray` benchmark from here https://github.com/dotnet/runtime/pull/61045#issuecomment-956006143 got additional +10% improvement