@EgorBo 2021-11-07
Next step is simple: fold
```asm
ubfiz   x4, x2, #2, #32
ldr     w4, [x1, x4]
```
into just
```asm
w4, [x1, x2, uxtw]
```
and we'll get a perfect codegen for cases when "arrRef + elemOffset" is CSE'd/hoisted.