@EgorBo 2021-11-08
With `LEA(BFIZ)` optimization:
```csharp
[Benchmark]
[ArgumentsSource(nameof(TestData))]
public int Sum(int[] array)
{
    int sum = 0;
    for (int i = 0; i < array.Length; i++)
    {
        sum += array[i];
    }
    return sum;
}
```
Benchmark results on Apple M1 arm64:
```
                             | Method |         array |     Mean |     Error |
                             |------- |-------------- |---------:|----------:|
Main without #61045 + #61035 |    Sum | Int32[100000] | 60,08 ns | 15.321 ns |
                Current Main |    Sum | Int32[100000] | 53.78 us |  0.007 us |
                     This PR |    Sum | Int32[100000] | 32.61 us |  0.082 us |
```
Codegen diff:
- Main vs This PR: https://www.diffchecker.com/4BPIsvbp
- Main before #61045 and #61035  vs This PR https://www.diffchecker.com/smwmMHvi