@EgorBo 2021-11-08
Diffs are much better when LEA(BFIZ) is enabled for short/char but there is an issue in the emitter (emits incorrect shift)