@EgorBo 2021-11-08
/azp run Fuzzlyn, runtime-coreclr outerloop