@EgorBo 2021-11-08
> @EgorBo - It might be good to post PerfScore diff as well, will be good to understand if there is lot of regression in it. Also, do you mind posting the diffs for benchmark collection?

How do I ask superpmi to generate perfscore btw?

> diffs for benchmark collection

it looks like it's not available for altjit mode I am running spmi under (I wasn't able to run spmi on my apple-m1)