@SingleAccretion 2021-11-08
> How do I ask superpmi to generate perfscore btw?

FWIW, I usually just `cd` into the directory SPMI created for diffs and `jit-analyze -b base -d diff -m PerfScore` directly.