@kunalspathak 2021-11-08
> How do I ask superpmi to generate perfscore btw?
> 
> >

It is only available for windows arm64.

```
D:\git\runtime>python src\coreclr\scripts\superpmi.py list-collections -target_arch arm64
================ Logging to E:\spmi\superpmi.364.log
Using JIT/EE Version from jiteeversionguid.h: 3df3e3ec-b1f6-4e6d-8439-2e7f3f7fa2ac
SuperPMI list-collections

5 collections for 3df3e3ec-b1f6-4e6d-8439-2e7f3f7fa2ac/windows/arm64/

https://clrjit2.blob.core.windows.net/superpmi/collections/3df3e3ec-b1f6-4e6d-8439-2e7f3f7fa2ac/windows/arm64/benchmarks.run.windows.arm64.checked.mch.zip
https://clrjit2.blob.core.windows.net/superpmi/collections/3df3e3ec-b1f6-4e6d-8439-2e7f3f7fa2ac/windows/arm64/coreclr_tests.pmi.windows.arm64.checked.mch.zip
https://clrjit2.blob.core.windows.net/superpmi/collections/3df3e3ec-b1f6-4e6d-8439-2e7f3f7fa2ac/windows/arm64/libraries.crossgen2.windows.arm64.checked.mch.zip
https://clrjit2.blob.core.windows.net/superpmi/collections/3df3e3ec-b1f6-4e6d-8439-2e7f3f7fa2ac/windows/arm64/libraries.pmi.windows.arm64.checked.mch.zip
https://clrjit2.blob.core.windows.net/superpmi/collections/3df3e3ec-b1f6-4e6d-8439-2e7f3f7fa2ac/windows/arm64/libraries_tests.pmi.windows.arm64.checked.mch.zip
```

You can use windows collection by passing `-target_os windows -target_arch arm64` to the `superpmi.py`.