@BruceForstall 2021-11-09
You should probably trigger jitstress, and also gcstress0xc to validate our assumption that GC will be happy with this new form of byref