@EgorBo 2021-11-09
Regression example:
```csharp
private static int[] _goto;

private static int Test(int destpos)
{
    _goto[destpos] = AddBacktrackNote();
    return _goto[destpos];
}
```
Here index is invariant but `_goto` arrayref is not (it can be changed in `AddBacktrackNote`) so in this case the default `"arrRef + (index + elemOffset)"` shape works better as we will be able to do CSE for `(index + elemOffset)`. Unfortunately I don't know how can I detect patterns like this early in morph. Naïve approaches like "if index is lvParam" don't really work well.
So I'm suggesting we merge it as is and look at regressions in dotnet/performance on perf triage meetings (if there will be a lot of them - address them or revert the commit completely).

A possible solution is a sort of tree re-association phase during CSE/LoopHoisting/VN e.g.
`(invarian1 + x) + invariant2` --reassoc--> `(invariant1 + invariant2) + x`